finished the code for producing simulation, you can see it in /Simulation/MakeSimulation.ipynb.
The simulation does not include sources, contains only diffusive gamma-ray galactic background, according to Fermi-LAT_pi0_map.npy. The resolution of background file is not very high, so when using healpy to generate the all-sky map and zoom it, there are huge pixels. But I think the data could already be sent to ML models and train them to de-convolution. De-convolution is only about the mapping between original image and image blurred by the psf. The ML model doesn't need to recognize the source and regard the source as an object.

To generate simulation data, simply run the MakeSimulation.ipynb, and the data will be saved as original.npy and blurred.npy in folder /SimulatedData/ .